server/requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.7
pillow==10.3.0
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.32.3
setuptools==70.0.0
sqlparse==0.4.4
urllib3==2.2.2
